Create a first implementation using Rc<RefCell<State>>. It sucks... fml.
The following is my current issue and the proposed solution which will
be implemented in the following commit.

Creating a graph is especially hard is rust because of the borrow
checker and the desire of immutability. The two possible ways I could
devise was to use a shit ton of Rc<RefCell<State>> so many states could
point to a single state and a state can be mutated (so we can connect
NFAs). The other problem with this solution is that RefCell cannot
really go into a HashSet/HashMap since Hash in rust seems to want to
hash based on data rather than memory location (or ptr value). The other
solution was to use unsafe rust, I didn't try implementing this since I
have a better solution now.

My current solution which is unimplemented is to simplify things.
Instead of passing around State structs and connecting them to each
other, which would lead to a single piece of data having many owners and
gets tedious with rusts borrow checker, I will just represent states by
integers. By doing this, instead of passing around ptrs which is quite
annoying in rust, we pass integers which point to a location in a
vector.

Lesson learned. Having just completed CS350 (intro to operating systems)
at UWaterloo, I have gotten into a rhythm of passing around pointers and
handling them on a very low level, even distinguishing between virtual
and physical pointers. Coding like this in rust is a path to despair I
think, handling ptrs like that is tough with the borrow checker. I think
at a fundamental level, I have to learn more idiomatic rust.

I don't usually write commit messages like diary entries, but
Rc<RefCell<>> everywhere had me very frustrated for this commit and I
wanted to write down my thoughts so I don't repeat this in the future.
I do like rust so far, but it can be frustrating.